adjust some data types

+++ features/transaction/cart/src/main/java/com/tokopedia/cart/view/CartListPresenter.kt
@@ -703,14 +703,14 @@ class CartListPresenter @Inject constructor(
 
         for (wholesalePriceData in wholesalePriceDataList) {
             if (itemQty >= wholesalePriceData.qtyMin) {
-                subTotalWholesalePrice = (itemQty * wholesalePriceData.prdPrc).toDouble()
+                subTotalWholesalePrice = (itemQty * wholesalePriceData.prdPrc)
                 hasCalculateWholesalePrice = true
                 val wholesalePriceFormatted = CurrencyFormatUtil.convertPriceValueToIdrFormat(
                     wholesalePriceData.prdPrc, false
                 ).removeDecimalSuffix()
                 cartItemHolderData.wholesalePriceFormatted = wholesalePriceFormatted
                 cartItemHolderData.wholesalePrice = wholesalePriceData.prdPrc
-                subtotalBeforeSlashedPrice = itemQty * cartItemHolderData.wholesalePrice.toDouble()
+                subtotalBeforeSlashedPrice = itemQty * cartItemHolderData.wholesalePrice
                 break
             }
         }
@@ -718,7 +718,7 @@ class CartListPresenter @Inject constructor(
         if (!hasCalculateWholesalePrice) {
             subTotalWholesalePrice = (itemQty * cartItemHolderData.productPrice)
             cartItemHolderData.wholesalePriceFormatted = null
-            cartItemHolderData.wholesalePrice = 0
+            cartItemHolderData.wholesalePrice = 0.0
             subtotalBeforeSlashedPrice =
                 if (cartItemHolderData.productOriginalPrice > 0) (itemQty * cartItemHolderData.productOriginalPrice)
                 else (itemQty * cartItemHolderData.productPrice)
